+++ src/App.tsx
@@ -47,6 +47,7 @@ import AdminNoticeAllPage from "./pages/admin/AdminNoticeAllPage";
 import AllReviewsPage from "./pages/spacedetail/AllReviewsPage"; // AllReviewsPage 임포트 추가
 import AdminBannerDetailPage from "./pages/admin/AdminBannerDetailPage";
 import BannerDetailPage from "./pages/mypage/BannerDetailPage";
+import AdminNoticeEditPage from "./pages/admin/AdminNoticeEditPage";
 
 const publicRoutes: RouteObject[] = [
   {
@@ -107,6 +108,7 @@ const protectedRoutes: RouteObject[] = [
       { path: "notices/:notificationId", element: <AdminNoticeDetailPage />, },
       { path: "notices/all", element: <AdminNoticeAllPage />, },
       { path: "banners/:bannerId", element: <AdminBannerDetailPage />, },
+      { path: "notices/edit/:notificationId", element: <AdminNoticeEditPage />, },
     ],
   },
 ];
